Fix Bearley (BER) (#23)
* Fix Bearley (BER)

The recording for Bearley (BER) was incorrectly marked as being for
Beauly (BEL). The latter appears to lack a recording entirely — I
suspect because it reopened in 2002 it might have missed the wave of
Phil Sayer recordings, and by then Scotland was already looking at using
its own announcer I suspect.

* chore: rename snippets

---------

Co-authored-by: GitHub Actions <41898282+github-actions[bot]@users.noreply.github.com>