Fix transcription of Fawley (#22)
* Fix transcription of Fawley

This sounds like an "f", is alphabetically in among South Western F
stations, and there have been various schemes mooted to reopen Fawley to
passengers so I strongly believe this is actually that. In addition
there is no such file in Celia's sound clips, plus there is already a
file for Horley which it was previously labelled as.

* chore: rename snippets

---------

Co-authored-by: GitHub Actions <41898282+github-actions[bot]@users.noreply.github.com>